Parses filtering issue: crash with empty dataset after filtering
**Filtering issue for "disabiquated" period `".@a"**  
Parses do not contain links between words and period (last '.' in a sentence). Links from ##LEFT-WALL### are ignored.  
In case of "disambiguated" period like `".@a"` this token is not parsed.  
If all sentences end with a "disambiguated" period, the filtered dataset in `kwargs['max_unparsed_words'] = 0` setting contains no parses.